[FIX] Change category property type from string to Category and removed unnecessary flows
