+++ app/src/main/java/com/architectcoders/bissu/data/database/entities/CategoryConverter.kt
@@ -3,21 +3,16 @@ package com.architectcoders.bissu.data.database.entities
 import androidx.room.TypeConverter
 import com.google.gson.Gson
 
-/**
- * Created by Anibal Cortez on 2020-01-19.
- */
 object CategoryConverter {
     @TypeConverter
     @JvmStatic
-    fun saveCategories(categories : List<Category>) : String{
-        val gson = Gson()
-        return gson.toJson(categories)
+    fun saveCategories(categorie: Category): String {
+        return Gson().toJson(categorie)
     }
 
     @TypeConverter
     @JvmStatic
-    fun getCategories(categories : String) : List<Category>{
-        val gson = Gson()
-        return gson.fromJson(categories, Array<Category>::class.java).toList()
+    fun getCategories(categories: String): Category {
+        return Gson().fromJson(categories, Category::class.java)
     }
 }
