+++ app/src/main/java/com/architectcoders/bissu/ui/book/BookDetailFragment.kt
@@ -180,7 +180,7 @@ class BookDetailFragment : Fragment() {
             bookdetail_bookpages.text = book.pages
             bookdetail_bookauthor.text = book.author
             bookdetail_bookeditorial.text = book.editorial
-            bookdetail_bookcategory.text = book.category
+            bookdetail_bookcategory.text = book.category.name
             bookdetail_bookdescription.text = book.description
         }
     }
