+++ data/src/main/java/com/architectcoders/data/repository/BookRepository.kt
@@ -4,9 +4,12 @@ import com.architectcoders.data.source.BookLocalDataSource
 import com.architectcoders.data.source.BookRemoteDatasource
 import com.architectcoders.domain.Book
 
-class BookRepository( private val localDataSource: BookLocalDataSource, private val remoteDatasource: BookRemoteDatasource) {
+class BookRepository(
+    private val localDataSource: BookLocalDataSource,
+    private val remoteDatasource: BookRemoteDatasource
+) {
 
-    suspend fun getBooks(): List<Book>  {
+    suspend fun getBooks(): List<Book> {
         if (localDataSource.isEmpty()) {
             remoteDatasource.getBooks().let {
                 localDataSource.saveBooks(it)
@@ -15,14 +18,7 @@ class BookRepository( private val localDataSource: BookLocalDataSource, private 
         return localDataSource.getBooks()
     }
 
-    suspend fun getBook(id : String) : Book {
-        if (localDataSource.isEmpty()) {
-            remoteDatasource.getBook(id).let {
-                it?.let {
-                    localDataSource.saveBooks(arrayListOf(it))
-                }
-            }
-        }
+    suspend fun getBook(id: String): Book {
         return localDataSource.getBook(id)
     }
 }
